Time: 2 ms (36.45%) | Memory: 11.2 MB (24.83%) - LeetSync

+++ 397-integer-replacement/integer-replacement.cpp
@@ -1,7 +1,7 @@
 class Solution {
-    unordered_map<long long, long long> mem;
+    unordered_map<long, int> mem;
 public:
-    int integerReplacement(long long n) {
+    int integerReplacement(long n) {
         if (n == 1) {
             return 0;
         }
